Works around emulated PCI devs not being recognized
For some reason out of my reach, when building a static library and linking
against it, emulated PCI devices were not being recognized by the kernel. So,
we are instead dumping the C files from XHyve so that CGO compiles then along
with the Go bindings. I tested again and after this change, emulated PCI
devices are recognized correctly by the Linux kernel.

We still need to remove calls to exit() from xhyve main thread so that we
can properly clean up resources from Go land.